Refactor package structure and add interface validations
