+++ Makefile
@@ -1,0 +1,12 @@
+run:
+	python main.py
+
+test:
+	pytest
+
+format:
+	black .
+
+lint:
+	ruff .
+
